HMAN-45 Refactored after initial code review
